Create README.md
Hi guys, I have here the last system project that I make when I’m a student. This M-Learning has lots of useful functions for you to use. This system also has a responsive design compatible with mobile devices. The system has 3 types of users which are the admin, teachers, and the students. The admin is in charge of maintaining important data such as the list subject, classes, department, and etc. This system can be also useful nowadays which some of the classes are under Modular Distance Learning. using this system the faculties or teacher can provide soft copies of learning documents to the students and also he/she can create a Practice Quiz and Assignment for a class. Students can answer the teacher's given practice quiz and limited to the teacher's allotted duration for taking the quiz. For the admin side of the system Add, edit, delete student information Add, edit, delete teacher information Add an entry like the subject, class, school calendar View log-in trail of every user View logs of every user activity For the Teacher User of the System The teacher can create his/her Class After creating a class the teacher View the students for that particular class Upload Downloadable materials for his/her class Create practice quizzes for students Grade students assignment Post announcements Create a class calendar of events Message co-teacher or students For the Student user of the system Sign up for his / her account Can view classmates in his / her class Answer practice quizzes View evaluation of assignment and quizzes Download downloadable materials in a class Upload downloadable material Message a teacher or his/ her classmate